feat: change styling to apply page
