+++ src/app/apply/[recruitmentCode]/_components/apply-narrative-box.tsx
@@ -2,28 +2,17 @@ import { useFormContext } from 'react-hook-form';
 import Container from '../../../../components/shared/container';
 import Typography from '../../../../components/shared/typography';
 import { IQuestion } from '../../../../lib/model/i-section';
-import { cn } from '../../../../lib/utils';
 import { IFormApplication } from '../page';
 
-interface ApplyNarrativeBoxProps {
-  question: IQuestion;
-  textareaProps?: React.TextareaHTMLAttributes<HTMLTextAreaElement>;
-}
-
-const ApplyNarrativeBox = ({ question }: ApplyNarrativeBoxProps) => {
-  const {
-    register,
-    watch,
-    setValue,
-    formState: { errors },
-  } = useFormContext<IFormApplication>();
+const ApplyNarrativeBox = ({ question }: { question: IQuestion }) => {
+  const { register, watch, setValue } = useFormContext<IFormApplication>();
 
   const currentAnswerIndex = watch('answers').findIndex(
     (answer) =>
       answer.questionId === question.id && answer.questionType === 'NARRATIVE',
   );
 
-  // make new answer if not exist, cuurentAnswerIndex === -1인 item 생성 방지를 위해 return null
+  // make new answer if not exist, currentAnswerIndex === -1인 item 생성 방지를 위해 return null
   if (currentAnswerIndex === -1) {
     setValue('answers', [
       ...watch('answers'),
@@ -46,29 +35,23 @@ const ApplyNarrativeBox = ({ question }: ApplyNarrativeBoxProps) => {
     ? `글자수 (${currentContent.length}/${question.wordLimit})`
     : '';
 
-  const isFieldError = errors.answers?.[currentAnswerIndex]?.content;
-
   const displayText = [necessityText, wordLimitText].filter(Boolean).join(', ');
 
   return (
-    <Container className="rounded-[0.625rem] bg-crews-w01 px-[1.25rem] py-[1.25rem]">
-      <div className="flex flex-col gap-[1rem]">
-        <div className="flex flex-col gap-[0.625rem]">
-          <Typography className="text-[1rem] font-bold text-crews-bk01">
+    <Container className="rounded-xl bg-crews-w01 p-3">
+      <div className="flex flex-col gap-4">
+        <div className="flex flex-col gap-1">
+          <Typography className="h-auto w-full text-sm font-semibold text-crews-bk01">
             {question.content}
           </Typography>
-          <Typography className="text-[0.875rem] text-crews-b06">
+          <Typography className="text-xs text-crews-b06">
             {displayText}
           </Typography>
         </div>
+
         <textarea
-          className={cn(
-            'w-full rounded-[0.625rem] p-[1rem] text-[0.875rem] outline outline-[1px] outline-crews-g04 placeholder:text-crews-g04',
-            {
-              'outline-crews-g04': !isFieldError,
-              'outline-crews-r03': isFieldError,
-            },
-          )}
+          rows={3}
+          className="w-full rounded-lg p-2 text-xs outline outline-1 outline-crews-g02 placeholder:font-light placeholder:text-crews-g03"
           placeholder="이곳에 답변을 입력해주세요."
           {...register(`answers.${currentAnswerIndex}.content`, {
             required: question.necessity
@@ -88,11 +71,6 @@ const ApplyNarrativeBox = ({ question }: ApplyNarrativeBoxProps) => {
           }}
         />
       </div>
-      {isFieldError && (
-        <Typography className="text-[0.875rem] text-crews-r03">
-          {errors.answers?.[currentAnswerIndex]?.content?.message}
-        </Typography>
-      )}
     </Container>
   );
 };
