+++ src/app/apply/[recruitmentCode]/_components/apply-selective-box.tsx
@@ -5,23 +5,15 @@ import { IQuestion } from '../../../../lib/model/i-section';
 import { IFormApplication } from '../page';
 import ApplyChoiceBox from './apply-choice-box';
 
-interface ApplySelectiveBoxProps {
-  question: IQuestion;
-}
-
-const ApplySelectiveBox = ({ question }: ApplySelectiveBoxProps) => {
-  const {
-    watch,
-    setValue,
-    formState: { errors },
-  } = useFormContext<IFormApplication>();
+const ApplySelectiveBox = ({ question }: { question: IQuestion }) => {
+  const { watch, setValue } = useFormContext<IFormApplication>();
 
   const currentAnswerIndex = watch('answers').findIndex(
     (answer) =>
       answer.questionId === question.id && answer.questionType === 'SELECTIVE',
   );
 
-  // make new answer if not exist, cuurentAnswerIndex === -1인 item 생성 방지를 위해 return null
+  // make new answer if not exist, currentAnswerIndex === -1인 item 생성 방지를 위해 return null
   if (currentAnswerIndex === -1) {
     setValue('answers', [
       ...watch('answers'),
@@ -51,17 +43,18 @@ const ApplySelectiveBox = ({ question }: ApplySelectiveBoxProps) => {
     .join(', ');
 
   return (
-    <Container className="rounded-[0.625rem] bg-crews-w01 px-[1.25rem] py-[1.25rem]">
-      <div className="flex flex-col gap-[1rem]">
-        <div className="flex flex-col gap-[0.625rem]">
-          <Typography className="text-[1.125rem] font-bold text-crews-bk01">
+    <Container className="rounded-xl bg-crews-w01 p-3">
+      <div className="flex flex-col gap-4">
+        <div className="flex flex-col gap-1">
+          <Typography className="h-auto w-full text-sm font-semibold text-crews-bk01">
             {question.content}
           </Typography>
-          <Typography className="text-[0.875rem] text-crews-b06">
+          <Typography className="text-xs text-crews-b06">
             {displayText}
           </Typography>
         </div>
-        <div className="flex flex-col gap-[0.5rem]">
+
+        <div className="flex flex-col gap-1">
           {question.choices.map((choice) => (
             <ApplyChoiceBox
               key={choice.id}
@@ -71,11 +64,6 @@ const ApplySelectiveBox = ({ question }: ApplySelectiveBoxProps) => {
           ))}
         </div>
       </div>
-      {errors.answers?.[currentAnswerIndex] && (
-        <Typography className="text-[0.875rem] text-crews-r03">
-          {errors.answers[currentAnswerIndex]?.message}
-        </Typography>
-      )}
     </Container>
   );
 };
