+++ src/app/apply/[recruitmentCode]/_components/header-section.tsx
@@ -10,7 +10,12 @@ const HeaderSection = () => {
         </div>
         <div className="flex flex-col gap-1 text-crews-g05">
           <p className="text-base font-light">
-            지원서는 모집 기간 이후 수정이 불가능하니 신중하게 작성해주세요.
+            지원서를 자유롭게 작성하세요. 모집 마감 전까지 자유롭게 제출
+            가능합니다.
+          </p>
+          <p className="text-base font-light">
+            모집 마감 이후에도 작성한 지원서를 확인하실 수 있지만 제출은 불가능
+            합니다 😃
           </p>
         </div>
       </div>
